Updated script names
